Query Helper Function/Member Profile API
added a function to return a list of dicts so that I didn't have to keep writing query and loop. Added a new api call to display all the rented movies a user has. Will need to tweak query so that I can get the film information but works ok for testing.